# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Oturum işlemleri",
                "description": "Oturum açma işlemi",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "mail@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "******"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Kullanıcı bilgisi",
                "operationId": "me",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Sipariş paket veya sipariş içeriklerini döner",
                "operationId": "b2fecac9dec0cdfa315b817cbb0066e0",
                "parameters": [
                    {
                        "name": "orderPackageId",
                        "in": "path",
                        "description": "Order Package UUID",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sipariş listesi hazırlandı"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Yeni sipariş oluştur",
                "operationId": "3e58ca2cf49e0b15f8f6b9719582b099",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "required": [
                                        "product_id",
                                        "amount"
                                    ],
                                    "properties": {
                                        "product_id": {
                                            "type": "string"
                                        },
                                        "amount": {
                                            "type": "number",
                                            "format": "decimal",
                                            "example": 0
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sipariş başarıyla oluşturuldu"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/discount/{orderPackageId}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "İndirim listesini döner",
                "operationId": "85eedfa9525dc109ccee48677598989e",
                "parameters": [
                    {
                        "name": "orderPackageId",
                        "in": "path",
                        "description": "Order Package UUID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sipariş listesi hazırlandı"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/orders/{orderPackageId}": {
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Siparişi iptal eder",
                "operationId": "5b3f836b5b204ec292b83428a0112369",
                "parameters": [
                    {
                        "name": "orderPackageId",
                        "in": "path",
                        "description": "Order Package UUID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sipariş paketi silindi"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/categories": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Kategoriler",
                "description": "Kategori listesini getirir",
                "operationId": "getCategories",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Ürün listesi",
                "description": "Returns a list of products",
                "operationId": "getProducts",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Category UUID",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Orders",
            "description": "Orders"
        },
        {
            "name": "Products",
            "description": "Products"
        }
    ]
}